+++ tests/examples/convert_jupyter_timeout/main.sh
@@ -1,0 +1,21 @@
+#!/usr/bin/env -S bash -x
+source ../example.rc
+
+# Run the example
+export PYDEVD_DISABLE_FILE_VALIDATION=1
+sb -w -j 1 & # > current_stdout.txt &
+PID=$!
+
+# Get the graph after completion of the pending steps.
+stepup wait
+stepup graph current_graph
+stepup join
+
+# Wait for background processes, if any.
+set +e; wait -fn $PID; RETURNCODE=$?; set -e
+[[ "${RETURNCODE}" -eq 0 ]] || exit 1
+
+# Check files that are expected to be present and/or missing.
+[[ -f slow.html ]] || exit 1
+[[ -f result.txt ]] || exit 1
+grep "done" result.txt